Shorts, shorts everywhere! This will lead to computers being loaded in a different state! Restart your computers! Sync active channel to client and draw indicator of active channel!

+++ src/main/java/li/cil/tis3d/api/detail/InfraredAPI.java
@@ -17,5 +17,5 @@ public interface InfraredAPI {
      * @param value     the value the packet carries.
      * @return the packet that was spawned.
      */
-    InfraredPacket sendPacket(final World world, final Vec3 position, final Vec3 direction, final int value);
+    InfraredPacket sendPacket(final World world, final Vec3 position, final Vec3 direction, final short value);
 }
