+++ src/main/java/li/cil/tis3d/common/api/InfraredAPIImpl.java
@@ -11,7 +11,7 @@ import net.minecraft.world.World;
  */
 public final class InfraredAPIImpl implements InfraredAPI {
     @Override
-    public InfraredPacket sendPacket(final World world, final Vec3 position, final Vec3 direction, final int value) {
+    public InfraredPacket sendPacket(final World world, final Vec3 position, final Vec3 direction, final short value) {
         final EntityInfraredPacket entity = new EntityInfraredPacket(world);
         entity.configure(position, direction.normalize(), value);
         world.spawnEntityInWorld(entity);
