+++ src/main/java/li/cil/tis3d/common/module/ModuleExecution.java
@@ -198,8 +198,8 @@ public final class ModuleExecution extends AbstractModuleRotatable {
             readFromNBT(nbt);
         } else {
             machine.getState().pc = nbt.getInteger(TAG_PC);
-            machine.getState().acc = nbt.getInteger(TAG_ACC);
-            machine.getState().bak = nbt.getInteger(TAG_BAK);
+            machine.getState().acc = nbt.getShort(TAG_ACC);
+            machine.getState().bak = nbt.getShort(TAG_BAK);
             if (nbt.hasKey(TAG_LAST)) {
                 try {
                     machine.getState().last = Optional.of(Enum.valueOf(Port.class, nbt.getString(TAG_LAST)));
@@ -326,8 +326,8 @@ public final class ModuleExecution extends AbstractModuleRotatable {
             writeToNBT(nbt);
         } else {
             nbt.setInteger(TAG_PC, machine.getState().pc);
-            nbt.setInteger(TAG_ACC, machine.getState().acc);
-            nbt.setInteger(TAG_BAK, machine.getState().bak);
+            nbt.setShort(TAG_ACC, machine.getState().acc);
+            nbt.setShort(TAG_BAK, machine.getState().bak);
             machine.getState().last.ifPresent(last -> nbt.setString(TAG_LAST, last.name()));
             nbt.setString(TAG_STATE, state.name());
         }
