+++ src/main/java/li/cil/tis3d/common/module/ModuleRandom.java
@@ -73,7 +73,7 @@ public final class ModuleRandom extends AbstractModule {
         final Pipe sendingPipe = getCasing().getSendingPipe(getFace(), port);
         if (!sendingPipe.isWriting()) {
             final Random random = getCasing().getCasingWorld().rand;
-            final int value = (short) random.nextInt(0xFFFF + 1);
+            final short value = (short) random.nextInt(0xFFFF + 1);
             sendingPipe.beginWrite(value);
         }
     }
