+++ src/main/java/li/cil/tis3d/common/module/execution/instruction/InstructionBitwiseAndImmediate.java
@@ -4,9 +4,9 @@ import li.cil.tis3d.common.module.execution.Machine;
 import li.cil.tis3d.common.module.execution.MachineState;
 
 public final class InstructionBitwiseAndImmediate implements Instruction {
-    private final int value;
+    private final short value;
 
-    public InstructionBitwiseAndImmediate(final int value) {
+    public InstructionBitwiseAndImmediate(final short value) {
         this.value = value;
     }
