+++ src/main/java/li/cil/tis3d/common/module/execution/instruction/InstructionBitwiseShiftRightImmediate.java
@@ -4,9 +4,9 @@ import li.cil.tis3d.common.module.execution.Machine;
 import li.cil.tis3d.common.module.execution.MachineState;
 
 public final class InstructionBitwiseShiftRightImmediate implements Instruction {
-    private final int value;
+    private final short value;
 
-    public InstructionBitwiseShiftRightImmediate(final int value) {
+    public InstructionBitwiseShiftRightImmediate(final short value) {
         this.value = value;
     }
